style: delete ensure.go
